refactor(ui): lazy-load narration controller off the startup chunk, unexport test-only symbols
